mounts-ab: Add slot switch support when mounting fails
In case of firmware mount failure, reboot and switch the slot,
if recovery boot is enabled.
If trial boot is enabled, then retry for max number of times
and then switch the slot if mount is still not successful.
In case of MTD devices, check for ubiattach return value and
fail_reboot if failure observed.
If ubiattach is successful but no volume available, fail_reboot in
that case as well.

Change-Id: I79ad18e5933eadfc043c4c5d4f0ee41ae71a35f0